refactor: split analysis job pool module
