+++ src/egui_app/controller/analysis_jobs/pool/job_progress.rs
@@ -1,0 +1,51 @@
+use super::db;
+use super::types::{AnalysisJobMessage, AnalysisProgress};
+use std::sync::{
+    Arc,
+    atomic::{AtomicBool, Ordering},
+    mpsc::Sender,
+};
+use std::thread::{JoinHandle, sleep};
+use std::time::Duration;
+
+#[cfg_attr(test, allow(dead_code))]
+pub(super) fn spawn_progress_poller(
+    tx: Sender<super::super::jobs::JobMessage>,
+    cancel: Arc<AtomicBool>,
+    shutdown: Arc<AtomicBool>,
+) -> JoinHandle<()> {
+    std::thread::spawn(move || {
+        let db_path = match super::library_db_path() {
+            Ok(path) => path,
+            Err(_) => return,
+        };
+        let conn = match db::open_library_db(&db_path) {
+            Ok(conn) => conn,
+            Err(_) => return,
+        };
+        let mut last: Option<AnalysisProgress> = None;
+        loop {
+            if shutdown.load(Ordering::Relaxed) {
+                break;
+            }
+            if cancel.load(Ordering::Relaxed) {
+                sleep(Duration::from_millis(200));
+                continue;
+            }
+            let progress = match db::current_progress(&conn) {
+                Ok(progress) => progress,
+                Err(_) => {
+                    sleep(Duration::from_millis(500));
+                    continue;
+                }
+            };
+            if last != Some(progress) {
+                last = Some(progress);
+                let _ = tx.send(super::super::jobs::JobMessage::Analysis(
+                    AnalysisJobMessage::Progress(progress),
+                ));
+            }
+            sleep(Duration::from_millis(200));
+        }
+    })
+}
